taming VAE

+++ initializations.py
@@ -7,5 +7,5 @@ def weight_variable_glorot(batch_size, input_dim, output_dim, name=""):
     """
     init_range = np.sqrt(6.0 / (input_dim + output_dim))
     initial = tf.random_uniform([batch_size, input_dim, output_dim], minval=-init_range,
-                                maxval=init_range, dtype=tf.float64)
+                                maxval=init_range, dtype=tf.float32)
     return tf.Variable(initial, name=name)
