+++ model.py
@@ -87,7 +87,7 @@ class GCNModelVAE(Model):
                                               act=tf.nn.relu,
                                               dropout=self.dropout,
                                               logging=self.logging)(self.inputs)
-        
+
         self.z_mean = GraphConvolution(batch_size=self.batch_size,
                                        input_dim=args.hidden_dim_1,
                                        output_dim=args.hidden_dim_2,
@@ -104,7 +104,7 @@ class GCNModelVAE(Model):
                                           dropout=self.dropout,
                                           logging=self.logging)(self.hidden1)
 
-        self.z = self.z_mean + tf.random_normal([self.n_samples, args.hidden_dim_2], dtype=tf.float64) * tf.exp(self.z_log_std)
+        self.z = self.z_mean + tf.random_normal([self.n_samples, args.hidden_dim_2]) * tf.exp(self.z_log_std)
 
         self.reconstructions = InnerProductDecoder(input_dim=args.hidden_dim_2,
                                       act=tf.nn.tanh,
