+++ preprocess.py
@@ -3,7 +3,7 @@ import tensorflow as tf
 
 def normalize_adj_tf(adj):
     rowsum = np.array(tf.reduce_sum(adj,2))
-    adj_norm = np.zeros(adj.shape, np.float64)
+    adj_norm = np.zeros(adj.shape, np.float32)
     for i in range(rowsum.shape[0]):
         degree_mat_inv_sqrt = np.diag(np.sign(rowsum[i])*np.power(np.abs(rowsum[i]), -0.5).flatten())
         degree_mat_inv_sqrt[np.isinf(degree_mat_inv_sqrt)] = 0.
@@ -12,7 +12,7 @@ def normalize_adj_tf(adj):
 
 def normalize_adj(adj):
     rowsum = np.array(adj.sum(2))
-    adj_norm = np.zeros(adj.shape, np.float64)
+    adj_norm = np.zeros(adj.shape, np.float32)
     for i in range(rowsum.shape[0]):
         degree_mat_inv_sqrt = np.diag(np.sign(rowsum[i])*np.power(np.abs(rowsum[i]), -0.5).flatten())
         degree_mat_inv_sqrt[np.isinf(degree_mat_inv_sqrt)] = 0.
